created and organized model folder. Login route working
