+++ src/model/db/departments.model.ts
@@ -1,0 +1,22 @@
+import { sequelize } from "src/sequelize/config.sequelize";
+import { INTEGER, ENUM, STRING, TEXT } from "sequelize";
+
+export const Departments = sequelize.define('department', {
+  id: {
+    type: INTEGER,
+    autoIncrement: true,
+    primaryKey: true,
+    allowNull: false,
+  },
+  name: {
+    type: TEXT,
+    allowNull: false,
+  },
+  description: {
+    type: TEXT,
+    allowNull: true,
+  },
+}, {
+  freezeTableName: true,
+  timestamps: false,
+});
