+++ src/model/db/index.ts
@@ -1,0 +1,3 @@
+export * from "./departments.model";
+export * from "./roles.model";
+export * from "./users.model";
