+++ src/model/index.ts
@@ -1,0 +1,5 @@
+// exportando index das databases
+export * from "./db";
+
+// 
+export * from "./login.model";
