+++ src/model/login.model.ts
@@ -1,84 +1,37 @@
-import { sequelize } from "src/sequelize/config.sequelize";
-import { INTEGER, ENUM, STRING, TEXT } from "sequelize";
-
-// decisao mais logica nesse caso é definir o modelo do banco na aqui model
-export const Users = sequelize.define('users', {
-  user_id: {
-    type: INTEGER,
-    autoIncrement: true,
-    primaryKey: true,
-    allowNull: false,
-  },
-  name: {
-    type: TEXT,
-    allowNull: false,
-  },
-  email: {
-    type: TEXT,
-    allowNull: false,
-  },
-  password: {
-    type: TEXT,
-    allowNull: false,
-  },
-  role: {
-    // type: ENUM('manager', 'analyst', 'admin'),
-    type: TEXT,
-    allowNull: true,
-  },
-  department_ID: {
-    type: INTEGER,
-    allowNull: true,
-  },
-}, {
-  timestamps: false,
-});
-
-export const Departments = sequelize.define('departments', {
-  department_ID: {
-    type: INTEGER,
-    autoIncrement: true,
-    primaryKey: true,
-    allowNull: false,
-  },
-  department_name: {
-    type: TEXT,
-    allowNull: false,
-  },
-  description: {
-    type: TEXT,
-    allowNull: false,
-  },
-}, {
-  timestamps: false,
-});
-
-Departments.hasMany(Users, {
-  foreignKey: 'department_ID',
-  onDelete: 'CASCADE',
-  onUpdate: 'CASCADE',
-});
+import { Departments } from "./db/departments.model";
+import { Roles } from "./db/roles.model";
+import { Users } from "./db/users.model";
 
 Users.belongsTo(Departments, {
-  foreignKey: 'department_ID',
-  onDelete: 'CASCADE',
-  onUpdate: 'CASCADE',
+  foreignKey: 'department_fk',
+  onDelete: 'NO ACTION',
+  onUpdate: 'NO ACTION',
+});
+
+Users.belongsTo(Roles, {
+  foreignKey: 'role_fk',
+  onDelete: 'NO ACTION',
+  onUpdate: 'NO ACTION',
 });
 
 class Department {
-  department_ID: number;
-  department_name: String;
+  id: number;
+  name: String;
   description: String;
 }
 
-export class User extends Department {
-  id_usuario: number;
+class Role {
+  id: number;
+  name: String;
+}
+
+export class User {
+  id: number;
   name: String;
   email: String;
   password: String;
-  tipo_usuario: ['manager', 'analyst', 'admin'];
-  department_ID: number;
   department: Department;
+  role: Role;
 }
 
 // planejo treansformar em um padrao para retorno de data para o front
